+++ client/src/actions_types/home/home_actions_types.js
@@ -7,4 +7,3 @@ export const UPDATE_SORT = 'UPDATE_SORT';
 export const FAILED_SEARCH = 'FAILED_SEARCH';
 export const GET_CATEGORIES = 'GET_CATEGORIES';
 export const CREATE_CATEGORY = 'CREATE_CATEGORY';
-export const GET_PRODUCT_REVIEW = 'GET_PRODUCT_REVIEW';
