+++ client/src/components/ProductDetailsCard/ProductDetailsTab/Reviews/UserReview/styles.js
@@ -1,0 +1,17 @@
+import { makeStyles } from "@material-ui/core/styles";
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    width: '100%',
+    backgroundColor: theme.palette.background.paper,
+    padding: 0,
+  },
+  inline: {
+    display: 'inline',
+  },
+  divider: {
+    width: '100%',
+  },
+}));
+
+export default useStyles;
