+++ client/src/components/ProductDetailsCard/ProductDetailsTab/Reviews/styles.js
@@ -1,0 +1,25 @@
+import { makeStyles } from "@material-ui/core/styles";
+
+const useStyles = makeStyles((theme) => ({
+    root: {
+      width: '100%',
+      maxWidth: 360,
+      backgroundColor: theme.palette.background.paper,
+      position: 'relative',
+      overflow: 'auto',
+      maxHeight: 185,
+      padding: 0,
+    },
+    listSection: {
+      backgroundColor: 'inherit',
+    },
+    ul: {
+      backgroundColor: 'inherit',
+      padding: 0,
+    },
+    notFound: {
+        textAlign: 'center',
+    }
+  }));
+
+export default useStyles;
